+++ internal/parse/parser.go
@@ -10,6 +10,7 @@ import (
 	"github.com/alexaandru/go-sitter-forest/cpp"
 	"github.com/alexaandru/go-sitter-forest/css"
 	"github.com/alexaandru/go-sitter-forest/dart"
+	"github.com/alexaandru/go-sitter-forest/elixir"
 	"github.com/alexaandru/go-sitter-forest/fish"
 	golang "github.com/alexaandru/go-sitter-forest/go"
 	"github.com/alexaandru/go-sitter-forest/hyprlang"
@@ -61,6 +62,7 @@ var grammars = map[string]func() unsafe.Pointer{
 	"php":        php.GetLanguage,
 	"kotlin":     kotlin.GetLanguage,
 	"dart":       dart.GetLanguage,
+	"elixir":     elixir.GetLanguage,
 	"csharp":     c_sharp.GetLanguage,
 }